requirements.txt
aiohttp
aiosignal
asgiref
async-timeout
boto3
botocore
celery
charset-normalizer
crypto
cryptocode
cryptography
Django==5.0.0
django-cors-headers
django-debug-toolbar
django-environ
django-extensions
django-filter
django-jalali
django-oauth-toolkit
django-oauth2-provider
django-redis
django-split-settings
django-url-filter
djangorestframework
djangorestframework-jwt
djangorestframework_simplejwt
djangorestframework-recursive
docopt
gunicorn
idna
importlib-metadata
jdatetime
jmespath
jwcrypto
multidict
oauth2-provider
oauthlib
openpyxl
packaging
pillow==9.5.0
prompt-toolkit
psycopg2
psycopg2-binary
pyasn1
pycparser
pycryptodomex
pyOpenSSL
pyparsing
python-dateutil
pytz
redis
requests
rsa==4.8
s3transfer
schedule
semantic-version
setuptools-rust
shellescape
shortuuid
six
sqlparse
tablib
typing-extensions
urllib3
XlsxWriter
zipp
django-simple-captcha
django-rest-captcha
pymemcache
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
django_elasticsearch_dsl_drf
django-crum
django-rest-swagger
drf-yasg
tinydb
django-simple-history
sentry_sdk
django-celery-beat
django-celery-results
channels
channels_redis
daphne
django-jazzmin